Color scales in the Brookings style

Color scales in the Brookings style

Usage

scale_color_brookings(
  palette = "brand1",
  discrete = TRUE,
  reverse = FALSE,
  ...
)

scale_fill_brookings(palette = "brand1", discrete = TRUE, reverse = FALSE, ...)

Arguments

palette

Character name of brookings_palettes

discrete

Boolean indicating whether color aesthetic is discrete or not

reverse

Boolean indicating whether the palette should be reversed

...

Additional arguments passed to discrete_scale() or scale_color_gradientn(), used respectively when discrete is TRUE or FALSE.

Palettes

Analogous

Different shades of the same hue, or of similar hues can be used when the associated values are related.

Contrasting

Colors on the opposite ends of the spectrum. Use Brookings Blue with Secondary colors.

Semantic

Where applicable, use colors that are associated with certain concepts. For e.g., semantic1, semantic2, and semantic3 could show subsets of gender data (female, male and other).

Positive & Negative

Shows pros, cons and neutral, or positive, negative and neutral data.

Political

Use red and blue of similar intensity to represent data related to political parties in the US. Yellow in political3 and political4 represents ‘Independent’ category

Categorical

Use categorical palettes to distinguish discrete categories of data that do not have an inherent ordering.

Sequential

Sequential palettes can be used to show an inherent order or variations in numeric values.

Diverging

Diverging palettes are useful when dealing with negative and positive values or a range of values that have two extremes with a baseline central value, like zero. The Brookings diverging palette should uses two distinct hues of similar brightness and saturation with a neutral color in the middle. Using a discrete set of colors with evenly distributed gradation can improve clarity of values relative to a continuous palette.

Misc

A pleasing option using Brookings Blue and accent yellow.